move AudioFileWriter to sdk/Audio
