+++ src/common.speech/DialogServiceTurnState.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import {AudioOutputFormatImpl} from "../sdk/Audio/AudioOutputFormat";
+import { AudioOutputFormatImpl } from "../sdk/Audio/AudioOutputFormat";
 import { AudioOutputStream, PullAudioOutputStreamImpl } from "../sdk/Audio/AudioOutputStream";
 import { DialogServiceTurnStateManager } from "./DialogServiceTurnStateManager";
 import { ActivityPayloadResponse, MessageDataStreamType } from "./ServiceMessages/ActivityResponsePayload";
